requirements.txt
fastapi
uvicorn
google-generativeai
python-telegram-bot
python-dotenv
numpy
psycopg2-binary # For pgvector
pydantic-settings